requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
openai-agents==0.0.7
python-multipart==0.0.9
numpy==1.26.4
sounddevice==0.4.6
pydub==0.25.1
websockets==12.0
python-dotenv==1.0.0

# Зависимости для тестирования
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0 